fix : displayOrder 데이터타입 int로 변경  (#35)

+++ src/main/java/com/kakaotech/team18/backend_server/domain/applicationFormField/entity/ApplicationFormField.java
@@ -31,7 +31,7 @@ public class ApplicationFormField extends BaseEntity {
     private boolean isRequired;
 
     @Column(name = "display_order")
-    private Long displayOrder;
+    private int displayOrder;
 
     @Convert(converter = StringListConverter.class)
     private List<String> options;
